feat: add @mina-js/klesia-utils package with handleJsonRpcRequest
Introduce a new package @mina-js/klesia-utils that exports the
handleJsonRpcRequest function. This utility converts JSON-RPC requests
into Mina node GraphQL requests, enabling seamless integration between
different request formats.
