+++ apps/klesia/package.json
@@ -18,19 +18,16 @@
 		"cleanup": "rimraf dist .turbo node_modules"
 	},
 	"devDependencies": {
+		"@mina-js/klesia-utils": "workspace:*",
 		"@mina-js/utils": "workspace:*"
 	},
 	"dependencies": {
 		"@hono/node-server": "1.13.8",
 		"@hono/zod-openapi": "0.19.2",
-		"@urql/core": "5.1.1",
-		"bigint-quantile": "0.0.2",
-		"dayjs": "1.11.13",
 		"dotenv": "16.4.7",
 		"hono": "4.7.4",
 		"hono-rate-limiter": "0.4.2",
 		"nanoid": "5.1.3",
-		"ofetch": "1.4.1",
 		"ts-pattern": "5.6.2",
 		"zod": "3.24.2"
 	}
