+++ packages/klesia-utils/package.json
@@ -1,7 +1,7 @@
 {
-	"name": "@mina-js/klesia-sdk",
-	"type": "module",
+	"name": "@mina-js/klesia-utils",
 	"version": "0.1.0",
+	"type": "module",
 	"module": "dist/index.js",
 	"types": "dist/index.d.ts",
 	"exports": {
@@ -13,18 +13,20 @@
 	"files": ["dist"],
 	"scripts": {
 		"build": "tsc --noEmit && tsup",
-		"test": "bun test --rerun-each 3",
+		"test": "bun test",
 		"cleanup": "rimraf dist .turbo"
 	},
-	"peerDependencies": {
-		"typescript": "^5.0.0"
+	"dependencies": {
+		"@urql/core": "5.1.1",
+		"mina-signer": "3.0.7",
+		"serialize-error": "12.0.0",
+		"superjson": "2.2.2"
 	},
 	"devDependencies": {
 		"@mina-js/utils": "workspace:*",
-		"@mina-js/klesia-utils": "workspace:*"
+		"zod": "3.24.2"
 	},
-	"dependencies": {
-		"micro-ftch": "0.4.2",
-		"ts-pattern": "5.6.2"
+	"peerDependencies": {
+		"typescript": "^5.0.0"
 	}
 }
