+++ packages/klesia-utils/src/node.ts
@@ -1,0 +1,8 @@
+import { Client, cacheExchange, fetchExchange } from "@urql/core";
+
+export const getNodeClient = (graphqlEndpoint: string) => {
+	return new Client({
+		url: graphqlEndpoint,
+		exchanges: [cacheExchange, fetchExchange],
+	});
+};
