+++ packages/klesia-utils/tsconfig.json
@@ -1,0 +1,3 @@
+{
+	"extends": "@tsconfig/bun/tsconfig.json"
+}
